+++ src/router/MainRouter.jsx
@@ -4,10 +4,8 @@ import { useDispatch, useSelector } from "react-redux";
 import { Navigate, Route, Routes } from "react-router-dom"
 import { LoginRegisterPage } from '../auth/pages/LoginRegisterPage';
 import { CheckingAuth } from "../components/CheckingAuth";
-import { FirebaseAuth } from "../firebase/config";
 import { useCheckOut } from "../hooks/useCheckout";
-import { MainPage } from "../pages/MainPage"
-import { login, logout } from "../store/auth/authSlice";
+import { MainPage } from "../pages/Main/MainPage";
 
 export const MainRouter = () => {
